Move RFC 4880 encoding to rfc4880.h
This allows making pgp_s2k.h internal in the future while leaving the
RFC 4880 logic available.
